chore: edit auth route layout component files
